refactor: move core logic out of the commands file
