+++ chores/utils.py
@@ -8,6 +8,7 @@ from selfservice.aggregator_adapter import get_aggregator_adapter
 from .models import Chore, ChoreVolunteer
 
 
+# LADEBUG: Fix loading from chores
 def get_chores_data(current_user_id=None, subset=None):
     aggregator_adapter = get_aggregator_adapter()
     if not aggregator_adapter:
@@ -29,7 +30,7 @@ def get_chores_data(current_user_id=None, subset=None):
         key = f"{turn.chore.id}-{turn.timestamp}"
         volunteers_by_key[key].append(turn.user)
 
-    # Drop use of aggregator virtualisation, internalise to CRM
+    # TODO: Drop use of aggregator virtualisation, internalise to CRM
     data = aggregator_adapter.get_chores()
     if data is None:
         return None, "No data available"
